+++ R/geom_pointgrid.R
@@ -37,9 +37,6 @@ geom_pointgrid <- function(
   show.legend = NA,
   inherit.aes = TRUE
 ) {
-  # input check
-  checkmate::assert_numeric(grid_x, any.missing = FALSE)
-  checkmate::assert_numeric(grid_y, any.missing = FALSE)
   # call layer function
   ggplot2::layer(
     mapping = mapping,
@@ -74,10 +71,14 @@ GeomPointGrid <- ggplot2::ggproto(
     if (is.character(data$shape)) {
       data$shape <- translate_shape_string(data$shape)
     }
-    # this line is the main difference to geom_point!
-    # the point coordinates are manipulated to map to a grid
-    # layout
-    data <- arrange_points_on_grid(data, grid_x, grid_y)
+    
+    # these lines are the main difference to geom_point!
+    # the point coordinates are manipulated to map to a grid layout
+    axes <- make_grid_axes_in_geom(data, grid_x, grid_y)
+    paog <- arrange_points_on_grid(axes, as.matrix(data[c("x", "y")]))
+    data[["x"]] <- paog[,1]
+    data[["y"]] <- paog[,2]
+    
     coords <- coord$transform(data, panel_params)
     ggname(
       "geom_pointgrid",
@@ -95,88 +96,38 @@ GeomPointGrid <- ggplot2::ggproto(
   }
 )
 
-arrange_points_on_grid <- function(tab, grid_x, grid_y) {
-  # define grid the input points should be mapped to
-  if (length(grid_x) == 1) {
-    if ("mapped_discrete" %in% class(tab[["x"]])) {
-      unique_x <- length(unique(tab[["x"]]))
-      x_grid <- seq(-0.5, unique_x+0.5, length.out = grid_x)
-    } else {
-      x_grid <- seq(min(tab[["x"]]), max(tab[["x"]]), length.out = grid_x)
-    }
+# prepare the axes coordinates from the grid input values
+make_grid_axes_in_geom <- function(tab, grid_x, grid_y) {
+  # input checks
+  checkmate::assert_data_frame(tab)
+  # compile axes
+                          # all, because it could be a vector
+  if (length(grid_x) == 1 & all(grid_x %% 1 == 0)) {
+    axis_x <- make_grid_sequence(
+      as.integer(grid_x), tab[["x"]],
+      ifelse(
+        "mapped_discrete" %in% class(tab[["x"]]),
+        "discrete",
+        "continuous"
+      )
+    )
   } else {
-    x_grid <- grid_x
+    axis_x <- grid_x
   }
-  if (length(grid_y) == 1) {
-    if ("mapped_discrete" %in% class(tab[["y"]])) {
-      unique_y <- length(unique(tab[["y"]]))
-      y_grid <- seq(-0.5, unique_y+0.5, length.out = grid_y)
-    } else {
-      y_grid <- seq(min(tab[["y"]]), max(tab[["y"]]), length.out = grid_y)
-    }
+  if (length(grid_y) == 1 & all(grid_y %% 1 == 0)) {
+    axis_y <- make_grid_sequence(
+      as.integer(grid_y), tab[["y"]],
+      ifelse(
+        "mapped_discrete" %in% class(tab[["y"]]),
+        "discrete",
+        "continuous"
+      )
+    )
   } else {
-    y_grid <- grid_y
+    axis_y <- grid_y
   }
-  xy_grid <- expand.grid(x_grid, y_grid)
-  # size check
-  if (nrow(xy_grid) < nrow(tab)) {
-    stop("The grid is not big enough to accommodate all input points.")
-  }
-  # calculate distance between input points and grid
-  distance_matrix <- fields::rdist(xy_grid, tab[c("x", "y")])
-  distance_long_raw <- stats::setNames(
-    reshape2::melt(distance_matrix), 
-    c('grid_id', 'point_id', 'distance')
-  )
-  distance_long <- distance_long_raw
-  # determine the optimal grid arrangement coordinates
-  grid_df <- core_arrange_loop(distance_long)
-  # replace input point coordinates with grid coordinates
-  tab[grid_df$point_id, c("x", "y")] <- xy_grid[grid_df$grid_id,]
-  return(tab)
-}
-
-# test code for the core algorithm
-# distance_matrix <- fields::rdist(expand.grid(1:5, 1:5), data.frame(x = c(2,2,3,3), y = c(2,2,3,3)))
-# distance_long <- stats::setNames(reshape2::melt(distance_matrix), c('grid_id', 'point_id', 'distance'))
-# core_arrange_loop(distance_long)
-# larger test
-# distance_matrix <- fields::rdist(expand.grid(1:100, 1:100), data.frame(x = sample(1:100, 4000, replace = T), y = sample(1:100, 4000, replace = T)))
-# distance_long <- stats::setNames(reshape2::melt(distance_matrix), c('grid_id', 'point_id', 'distance'))
-# core_arrange_loop(distance_long)
-
-core_arrange_loop <- function(distance_long) {
-  distance_long_sorted <- dplyr::arrange(distance_long, .data[["distance"]])
-  res_total <- list()
-  repeat {
-    res <- arrange_iteration(distance_long_sorted)
-    res_total <- append(res_total, list(res$solved))
-    if (nrow(res$todo) > 0) {
-      distance_long_sorted <- res$todo
-    } else {
-      break
-    }
-  }
-  return(dplyr::bind_rows(res_total))
-}
-
-arrange_iteration <- function(distance_long_sorted) {
-  # get smallest distance grid point for each input point
-  closest_grid_points <- distance_long_sorted[!duplicated(distance_long_sorted$point_id),]
-  # find grid point duplicates and create subsets with already uniquely claimed positions
-  grid_points_dups <- unique(closest_grid_points$grid_id[duplicated(closest_grid_points$grid_id)])
-  has_grid_point_dup <- closest_grid_points$grid_id %in% grid_points_dups
-  without_dups <- closest_grid_points[!has_grid_point_dup,]
-  with_dups <- closest_grid_points[has_grid_point_dup,]
-  # find best input point matches according to distance for duplicated grid points
-  with_dups_better <- with_dups[!duplicated(with_dups$grid_id),]
-  # combine newly solved point-grid pairs for the output grid
-  newly_solved <- dplyr::bind_rows(without_dups, with_dups_better)
-  # construct distance table with yet unattributed input and grid points
-  new_distance_long <- dplyr::anti_join(
-    dplyr::anti_join(distance_long_sorted, newly_solved, by = "grid_id"),
-    newly_solved, by = "point_id"
-  )
-  # output
-  return(list(solved = newly_solved, todo = new_distance_long))
+  # expand
+  axes_df <- expand.grid(axis_x, axis_y)
+  # return axes matrix
+  return(as.matrix(axes_df))
 }