+++ playground/example.R
@@ -1,40 +1,52 @@
-#### simple example ####
-
 library(ggplot2)
 
+#### generate test data ####
+
+n <- 3000
 my_data <- tibble::tibble(
-  x = rnorm(500),
-  y = rnorm(500),
-  value = sample(c("A", "B"), 500, replace = T)
+  x = runif(n, -5, 5),
+  y = runif(n, -5, 5),
+  value = sample(c("A", "B"), n, replace = T),
+  id = 1:n
 )
 
-p1 <- ggplot(my_data, aes(x = x, y = y, color = value)) +
+#### normal plotting test ####
+
+ggplot(my_data, aes(x = x, y = y, color = value)) +
   geom_point() +
-  coord_cartesian(xlim = c(-5, 5), ylim = c(-5, 5))
+  coord_fixed(xlim = c(-5, 5), ylim = c(-5, 5))
 
-p2 <- ggplot(my_data, aes(x = x, y = y, color = value)) +
-  ggpointgrid::geom_pointgrid(grid_x = 50, grid_y = 50) +
-  coord_cartesian(xlim = c(-5, 5), ylim = c(-5, 5))
-
-cowplot::plot_grid(p1, p2)
-
+ggplot(my_data, aes(x = x, y = y, color = value)) +
+  ggpointgrid::geom_pointgrid(grid_x = 80, grid_y = 80) +
+  coord_fixed(xlim = c(-5, 5), ylim = c(-5, 5))
 
 #### algorithm test ####
 
-# generate test data
-regular_grid <- expand.grid(1:100, 1:100)
-points_to_plot <- data.frame(
-  x = sample(1:100, 4000, replace = T),
-  y = sample(1:100, 4000, replace = T)
+axis_x <- make_grid_sequence(80L, my_data$x)
+axis_y <- make_grid_sequence(80L, my_data$y)
+axes <- expand.grid(axis_x, axis_y)
+
+system.time(
+  res1 <- arrange_points_on_grid(
+      as.matrix(axes),
+      as.matrix(my_data[c("x", "y")])
+    )
 )
 
-# calculate pairwise distances of grid and points
-distance_matrix <- fields::rdist(regular_grid, points_to_plot)
-distance_long <- stats::setNames(
-  reshape2::melt(distance_matrix),
-  c("grid_id", "point_id", "distance")
+#### futhark debugging ####
+
+test_data_string <- paste(
+  "[", paste(axes[[1]], collapse = ","), "]",
+  "[", paste(axes[[2]], collapse = ","), "]",
+  "[", paste(my_data$x, collapse = ","), "]",
+  "[", paste(my_data$y, collapse = ","), "]"
 )
+cli_cmd <- paste("echo", test_data_string, "| ./src/arrange")
 
-# run algorithm
-core_algorithm(distance_long)
+# run arrange on command line - needs to be compiled as an executable first
+result <- system(cli_cmd, intern = T)
+system(paste("time", cli_cmd))
 
+# write test data string to file to copy it into the profiling setup list
+# in arrange.fut
+writeLines(test_data_string, "futhark_test.txt")